docs: Add note to install Testify
The Hugo Contributor Guide (Development) instructs users to run "go test ./..." and ensure it passes before moving on. However Hugo requires the Testify package for testing, and running the go test command without Testify results in 'cannot find package' errors and failing tests. It's hard to understand what to do next for users not familiar with Go dependencies. This commit adds a note to the contributor guide instructing users to get Testify if they don't already have it installed. Running the 'go get' command added here results in passing tests in the next step of the contributor guide.