Remove delimiters in YAML and TOML examples
The use of `---` and `...` delimiters in the YAML examples seem to
suggest that those delimiters are necessary in a .yaml/.yml file when in
fact they are optional. Also, it seems to be common practice to *not*
use `---` and `...` in .yaml/.yml files. Similarly, it's less common to
use `+++` in .toml files. The examples on TOML's Github repo don't use
it, and Hugo-generated .toml files don't use it either.  This commit
removes the delimiters mentioned above to conform with the more common
styles.

Additionally this modifies the example for shortcut to find
configuration option

- add `cd` to make it clear that user has to change to the hugo site's
  directory first
- separate the command and the output of the command