Update quickstart
Quickstart step 4 makes reference to outdated instructions (https://gohugo.io/getting-started/quick-start/#step-4-add-some-content).

As it shows the command output, it used to build a draft by default, now it is generating a regular post with empty values. To make the post display something when running the server, example content should be added.

Also there is no need to start the server with the `-D` option as new there isn't any draft, and `date` has an example value, but if none is specified it will just ignore it displaying an error in console.