Add tip about clearing browser cache
A forum user expressed confusion about why the wrong site was being served; it turned out he had not cleared his cache, and the old site was showing up.  I have proposed an addition to the last tip about seeing browser changes right away (in the "Site Configuration" section) that suggests you may need to clear the cache and provides a link to a page detailing how to do it.