Replace javascript:void(0) docs menu toggle links
This commit replaces links in the docs menu from javascript:void(0)
to links to the respective overview pages, and disables these links
in javascript (retaining their function as dropdown toggles).

These links to the overview pages to be used by actions and
interfaces outside of regular left clicks, such as through a middle
mouse click to open a new tab, or interacting with the link via the
context menu (right click).

As a positive side-effect, this change also allows for the docs menu
to be navigable by those with javascript disabled.