Fix ToC example to use binary true/false
The example code seemed to require that `toc` be set to the *string* `"false"` instead of the boolean keyword `false`. However, the text said the `toc` variable should be set to boolean `true`/`false` keywords, not strings. I'm not very familiar with GoLang so I could be completely wrong. I just know that the following Hugo page preamble did not work with the original code, but does work with the proposed fix.

```
+++

toc = false

+++
```