Remove purge from Webpack, temporarily
Purge is removing CSS from the themes site because that site is not present for Purge to know to keep those classes. This change updates the CSS build to not use Purge to fix the currently broken public site, but commenting out only to add back when we have a permanent solution.