show 'description' when hovered on function nav link
### Changes
Added `title` attribute with function's description as its value, so that user hovering an entry of a function in sidebar (nav links) sees the function's short description. Attribute is added to `<a>` tag, alternatively it could be added to `<li>` tag too; though it wouldn't make any major difference. Closes #999.
### Previously
```
line 13
            <a href="{{.URL}}" class="db link hover-bg-gray hover-white pl3 pr2 pv2 {{if $currentPage.IsMenuCurrent "docs" . }}primary-color {{ else }}black {{end}}">
```
### Now
```
line 13
            <a href="{{.URL}}" class="db link hover-bg-gray hover-white pl3 pr2 pv2 {{if $currentPage.IsMenuCurrent "docs" . }}primary-color {{ else }}black {{end}}"{{if .description}} title="{{.description}}"{{end}}>
```