fix new post command directory
The quickstart says that hugo new posts/my-first-post.md is the command to create new posts but it wasn't working. I could see another person with the same problem here: https://stackoverflow.com/questions/51974423/hugo-posts-not-displaying. It seems that by default hugo is looking for posts in content/posts/ not posts/