Correct possible typos re: Alias robots usage
If the `<meta name="robots" content="noindex">` is on the old alias page, `https://example.com/posts/my-old-url/`, then the directive is informing the robot to not index the old page. I've updated this to clarify.

I also removed "crawl" as it implies using "no-index" tells a search engine bot crawl the page for additional links. Unless the `nofollow` directive is included search engine bots should still look through the page for links.

[Google's clarification on `robots` directives:](https://developers.google.com/search/reference/robots_meta_tag)
`noindex` = "Do not show this page in search results." 
`nofollow` = "Do not follow the links on this page."