Update page-bundles.md
In the first table under Branch Bundle, the Allowed Resources are "Only non-page (like images, pdf, etc.) types".  
But is this true?    In the example at the bottom of the page: "branch-bundle-1
This branch bundle has the _index.md, two other content Markdown files and two image files.".  Thank you.