Clean documentation summary on CLI commands page
Lack of 'description' frontmatter generated by spf13/cobra
resulted in ugly summary pages. This patch fixes that by
inserting itself before the default summary use with a
custom code snippet extracting the synopsis as summary.

There is no direct access to the content DOM so
string matching is used to identify the start and end
of the synopsis section.