Remove my defunct and little used migrator
wpxr-to-static is not much used, and I have had recent reports that it
does not work with current wordpress.com exports. As I have not needed
to update the project myself and it is little used, I remove it from the
Hugo documentation as I will be marking it archived and unmaintained.

Signed-off-by: Daniel F. Dickinson <dfdpublic@wildtechgarden.ca>