theme: Generate QR code once per page
We currently generate two QR codes for each page, one in the header and
another at the top the printed page. The content of the two codes is
identical; the only difference is scale.

Instead of generating two images, make the same function call in both
places so that the second call uses the cached instance.